Refactor to an api client library.
